pyproject.toml
[tool.poetry]
name = "crime-dashboard"
version = "0.1.0"
description = ""
authors = ["Arkadiy Sotnikov  <arkadiy.sotnikov@burtcorp.com>"]

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
pandas = "^1.5.0"
openpyxl = "^3.0.10"
geopandas = "^0.11.1"
requests = "^2.28.1"
hydra-core = "^1.2.0"
torchvision = "^0.13.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

